apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "knowledge-rss-api"
version = "0.1.0"
description = "FastAPI backend for RSS aggregation, hot ranking, and AI summaries"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "apscheduler>=3.11.0",
  "asyncpg>=0.30.0",
  "fastapi>=0.115.0",
  "feedparser>=6.0.11",
  "httpx>=0.28.1",
  "pydantic>=2.10.6",
  "pydantic-settings>=2.8.1",
  "pyjwt[crypto]>=2.10.1",
  "python-dateutil>=2.9.0.post0",
  "uvicorn[standard]>=0.34.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5",
  "pytest-asyncio>=0.25.3",
  "ruff>=0.11.4"
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
